+++ libfreenect/cppbuild.sh
@@ -7,7 +7,7 @@ if [[ -z "$PLATFORM" ]]; then
     exit
 fi
 
-LIBFREENECT_VERSION=0.5.1
+LIBFREENECT_VERSION=0.5.2
 download https://github.com/OpenKinect/libfreenect/archive/v$LIBFREENECT_VERSION.zip libfreenect-$LIBFREENECT_VERSION.zip
 
 mkdir -p $PLATFORM